+++ frontend/components/BoardCarousel.tsx
@@ -1,7 +1,15 @@
 import { faArrowLeft, faArrowRight } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { useState, forwardRef, Dispatch, SetStateAction, useEffect, Children, ReactNode, FC, useRef, createRef, useLayoutEffect } from 'react';
-import FlipMove from 'react-flip-move';
+import {
+  useState,
+  forwardRef,
+  Dispatch,
+  SetStateAction,
+  Children,
+  createRef,
+  useLayoutEffect,
+} from 'react';
+import { usePrevious } from '../hooks';
 import boardData from '../public/boardData.json';
 
 const getStyle = (index: number): string => {
@@ -68,20 +76,10 @@ const rotateLeft = (
   setBoard(temp);
 };
 
-function usePrevious(value: any) {
-  const ref = useRef();
-
-  useEffect(() => {
-    ref.current = value;
-  }, [value]);
-
-  return ref.current;
-}
-
 const calculateBoundingBoxes = (children: any) => {
   const boundingBoxes: any = {};
 
-  Children.forEach(children, child => {
+  Children.forEach(children, (child) => {
     const domNode = child.ref.current;
     const nodeBoundingBox = domNode.getBoundingClientRect();
 
@@ -91,7 +89,7 @@ const calculateBoundingBoxes = (children: any) => {
   return boundingBoxes;
 };
 
-const AnimateChildren = ({children}: {children: any}) => {
+const AnimateChildren = ({ children }: { children: any }) => {
   const [boundingBox, setBoundingBox]: any = useState({});
   const [prevBoundingBox, setPrevBoundingBox]: any = useState({});
   const prevChildren = usePrevious(children);
@@ -110,7 +108,7 @@ const AnimateChildren = ({children}: {children: any}) => {
     const hasPrevBoundingBox = Object.keys(prevBoundingBox).length;
 
     if (hasPrevBoundingBox) {
-      Children.forEach(children, child => {
+      Children.forEach(children, (child) => {
         // TODO: Add support for more than 5 members
         const domNode = child.ref.current;
         const firstBox = prevBoundingBox[child.key];
@@ -121,11 +119,11 @@ const AnimateChildren = ({children}: {children: any}) => {
         if (changeInX) {
           requestAnimationFrame(() => {
             domNode.style.transform = `translateY(${changeInY}px) translateX(${changeInX}px)`;
-            domNode.style.transition = "transform 0s";
+            domNode.style.transition = 'transform 0s';
 
             requestAnimationFrame(() => {
-              domNode.style.transform = "";
-              domNode.style.transition = "transform 500ms";
+              domNode.style.transform = '';
+              domNode.style.transition = 'transform 500ms';
             });
           });
         }
@@ -134,7 +132,7 @@ const AnimateChildren = ({children}: {children: any}) => {
   }, [boundingBox, prevBoundingBox, children]);
 
   return children;
-}
+};
 
 const BoardCarousel = () => {
   const [board, setBoard] = useState(boardData);
@@ -144,23 +142,28 @@ const BoardCarousel = () => {
       {/* FIXME: https://github.com/joshwcomeau/react-flip-move/issues/273 */}
       {/* <FlipMove className='grid grid-cols-5 gap-10'> */}
       <AnimateChildren>
-      {board.slice(0, 5).map((value, index) => {
-        if (index === 2) {
+        {board.slice(0, 5).map((value, index) => {
+          if (index === 2) {
+            return (
+              <MiddleBoardCarouselItem
+                key={value.key}
+                index={index}
+                value={value}
+                board={board}
+                setBoard={setBoard}
+                ref={createRef()}
+              />
+            );
+          }
           return (
-            <MiddleBoardCarouselItem
+            <BoardCarouselItem
               key={value.key}
               index={index}
               value={value}
-              board={board}
-              setBoard={setBoard}
               ref={createRef()}
             />
           );
-        }
-        return (
-          <BoardCarouselItem key={value.key} index={index} value={value} ref={createRef()} />
-        );
-      })}
+        })}
       </AnimateChildren>
       {/* </FlipMove> */}
     </div>
@@ -183,15 +186,17 @@ const MiddleBoardCarouselItem = forwardRef<HTMLDivElement, MiddleProps>(
             <FontAwesomeIcon size='lg' icon={faArrowLeft}></FontAwesomeIcon>
           </button>
 
-          <div ref={ref} key={props.value.key} className={`avatar ${getPlacement(props.index)}`}>
+          <div
+            ref={ref}
+            key={props.value.key}
+            className={`avatar ${getPlacement(props.index)}`}
+          >
             <div
               className={`
                    mask mask-squircle ${getStyle(props.index)}`}
             >
               <img alt='e' src={props.value.url} />
-              <div>
-
-              </div>
+              <div></div>
             </div>
           </div>
